+++ lib/vagrant/action/vm/host_name.rb
@@ -2,13 +2,13 @@ module Vagrant
   class Action
     module VM
       class HostName
-        def initialize(app, env, options=nil)
+        def initialize(app, env)
           @app = app
-          env.merge!(options || {})
         end
 
         def call(env)
           @app.call(env)
+
           host_name = env["config"].vm.host_name
           if !host_name.nil?
             env.ui.info I18n.t("vagrant.actions.vm.host_name.setting")