+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   - Change error output with references to VirtualBox 3.2 to 4.0.
   - Internal SSH through net-ssh now uses `IdentitiesOnly` thanks to
     upstream net-ssh fix.
+  - Fix issue causing warnings to show with `forwardx11` enabled for SSH. [GH-279]
 
 ## 0.7.0 (January 19, 2011)