+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.7.1 (unreleased)
+## 0.7.1 (January 28, 2011)
 
   - Change error output with references to VirtualBox 3.2 to 4.0.
   - Internal SSH through net-ssh now uses `IdentitiesOnly` thanks to